+++ src/probemcp/mi/__init__.py
@@ -1,0 +1,7 @@
+"""GDB/MI parsing primitives."""
+
+from probemcp.mi.errors import MIParseError
+from probemcp.mi.parser import parse_mi_record
+from probemcp.mi.records import MIRecord, MIRecordKind, MIStreamRecord
+
+__all__ = ["MIParseError", "MIRecord", "MIRecordKind", "MIStreamRecord", "parse_mi_record"]